\name{LAD_list}
\alias{LAD_list}
\docType{data}
\title{a summary of 19 common LAD models}
\description{
  \code{LAD_list} is a data frame comprising 10 common leaf angle distributions models that are considered in the hemiphoto2LAI package.
}
\usage{
    data(LAD_list)
}
\source{
hemiphoto2LAI v0.1
}
 
\examples{
library(hemiphoto2LAI)

\donttest{
data(LAD_list)
}
}
\keyword{misc}
