\name{LAD_list}
\alias{LAD_list}
\docType{data}
\title{A summary of 19 common LAD models}
\description{
  \code{LAD_list} is a data frame comprising 10 common leaf angle distributions models that are considered in the hemiphoto2LAI package.
}
\usage{
    data(LAD_list)
}
\value{
   See Zheo et al. (2019) to get more information about these 19 LAD models, especially their exact equaiton forms.
   }
\source{
hemiphoto2LAI v0.1
}
 
\examples{
library(hemiphoto2LAI)

\donttest{
data(LAD_list)
}
}
\references{
Zhao et al. (2019). How to better estimate leaf area index and leaf angle distribution from digital hemispherical photography? Switching to a binary nonlinear regression paradigm (under review) 
}
\keyword{misc}
